#JC-954 Secured removal of section & component in the same was as branches (poulpe should pass admin password)

+++ jcommune-view/jcommune-web-controller/src/main/java/org/jtalks/jcommune/web/controller/integration/PoulpeNotificationHandler.java
@@ -69,21 +69,6 @@ public class PoulpeNotificationHandler {
     }
 
     /**
-     * Checks whether the admin password is not blank and it matches the admin password from database. Note that Poulpe
-     * sends hash of the password to us and thus we compare it with has as well.
-     *
-     * @param adminPassword the password sent by Poulpe
-     * @throws IllegalArgumentException if the password sent by Poulpe is blank or does not match the one in the
-     *                                  database
-     */
-    private void assertAdminPasswordCorrect(String adminPassword) throws NotFoundException {
-        checkArgument(isNotBlank(adminPassword), "No password specified while it is required");
-        User admin = userService.getCommonUserByUsername("admin");
-        checkArgument(adminPassword.equals(admin.getPassword()),
-                "Wrong password was specified during removal of branch/section/component.");
-    }
-
-    /**
      * Handles notification about section deletion. Removes all topics from section. Sections and branches won't be
      * removed
      *
@@ -92,7 +77,9 @@ public class PoulpeNotificationHandler {
      */
     @RequestMapping(value = "/sections/{sectionId}", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteSection(@PathVariable("sectionId") long sectionId) throws NotFoundException {
+    public void deleteSection(@PathVariable("sectionId") long sectionId,
+                              @RequestParam(value = "password") String adminPassword) throws NotFoundException {
+        assertAdminPasswordCorrect(adminPassword);
         sectionService.deleteAllTopicsInSection(sectionId);
     }
 
@@ -104,7 +91,8 @@ public class PoulpeNotificationHandler {
      */
     @RequestMapping(value = "/component", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteComponent() throws NotFoundException {
+    public void deleteComponent(@RequestParam(value = "password") String adminPassword) throws NotFoundException {
+        assertAdminPasswordCorrect(adminPassword);
         sectionService.deleteAllTopicsInForum();
     }
 
@@ -122,6 +110,21 @@ public class PoulpeNotificationHandler {
         ModelAndView mav = new ModelAndView(jsonView);
         mav.addObject(ERROR_MESSAGE_PARAMETER, exception.getMessage());
         return mav;
+    }
+
+    /**
+     * Checks whether the admin password is not blank and it matches the admin password from database. Note that Poulpe
+     * sends hash of the password to us and thus we compare it with has as well.
+     *
+     * @param adminPassword the password sent by Poulpe
+     * @throws IllegalArgumentException if the password sent by Poulpe is blank or does not match the one in the
+     *                                  database
+     */
+    private void assertAdminPasswordCorrect(String adminPassword) throws NotFoundException {
+        checkArgument(isNotBlank(adminPassword), "No password specified while it is required");
+        User admin = userService.getCommonUserByUsername("admin");
+        checkArgument(adminPassword.equals(admin.getPassword()),
+                "Wrong password was specified during removal of branch/section/component.");
     }
 
 }
